// web/src/api/models/IptoolsSubnetInfo.ts
export interface IptoolsSubnetInfo {
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    address: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    address_bits: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    broadcast?: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    broadcast_bits?: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    _class?: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    family: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    host_max: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    host_max_bits: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    host_min: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    host_min_bits: string;
    /**
     * 
     * @type {boolean}
     * @memberof IptoolsSubnetInfo
     */
    host_route?: boolean;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    hosts: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    netmask: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    netmask_bits: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    network: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    network_bits: string;
    /**
     * 
     * @type {number}
     * @memberof IptoolsSubnetInfo
     */
    prefix: number;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    scope: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    wildcard: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsSubnetInfo
     */
    wildcard_bits: string;
}