// web/src/api/models/IptoolsReverseDNS.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IptoolsReverseDNS
 */
export interface IptoolsReverseDNS {
    /**
     * 
     * @type {string}
     * @memberof IptoolsReverseDNS
     */
    family: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsReverseDNS
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof IptoolsReverseDNS
     */
    zone?: string;
}

/**
 * Check if a given object implements the IptoolsReverseDNS interface.
 */
export function instanceOfIptoolsReverseDNS(value: object): value is IptoolsReverseDNS {
    if (!('family' in value) || value['family'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function IptoolsReverseDNSFromJSON(json: any): IptoolsReverseDNS {
    return IptoolsReverseDNSFromJSONTyped(json, false);
}

export function IptoolsReverseDNSFromJSONTyped(json: any, ignoreDiscriminator: boolean): IptoolsReverseDNS {
    if (json == null) {
        return json;
    }
    return {
        
        'family': json['family'],
        'name': json['name'],
        'zone': json['zone'] == null ? undefined : json['zone'],
    };
}

export function IptoolsReverseDNSToJSON(json: any): IptoolsReverseDNS {
    return IptoolsReverseDNSToJSONTyped(json, false);
}

export function IptoolsReverseDNSToJSONTyped(value?: IptoolsReverseDNS | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'family': value['family'],
        'name': value['name'],
        'zone': value['zone'],
    };
}