// web/src/api/models/IptoolsRangeCIDRs.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IptoolsRangeCIDRs
 */
export interface IptoolsRangeCIDRs {
    /**
     * 
     * @type {Array<string>}
     * @memberof IptoolsRangeCIDRs
     */
    cidrs: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof IptoolsRangeCIDRs
     */
    family: string;
}

/**
 * Check if a given object implements the IptoolsRangeCIDRs interface.
 */
export function instanceOfIptoolsRangeCIDRs(value: object): value is IptoolsRangeCIDRs {
    if (!('cidrs' in value) || value['cidrs'] === undefined) return false;
    if (!('family' in value) || value['family'] === undefined) return false;
    return true;
}

export function IptoolsRangeCIDRsFromJSON(json: any): IptoolsRangeCIDRs {
    return IptoolsRangeCIDRsFromJSONTyped(json, false);
}

export function IptoolsRangeCIDRsFromJSONTyped(json: any, ignoreDiscriminator: boolean): IptoolsRangeCIDRs {
    if (json == null) {
        return json;
    }
    return {
        
        'cidrs': json['cidrs'],
        'family': json['family'],
    };
}

export function IptoolsRangeCIDRsToJSON(json: any): IptoolsRangeCIDRs {
    return IptoolsRangeCIDRsToJSONTyped(json, false);
}

export function IptoolsRangeCIDRsToJSONTyped(value?: IptoolsRangeCIDRs | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cidrs': value['cidrs'],
        'family': value['family'],
    };
}